+++ changelog.md
@@ -13,6 +13,16 @@ will not have contiguous patch numbers. Initial major and minor releases will be
 in this file without a patch number. Patch version will be included for bug fix releases, but
 may not exactly match a publicly released version.
 
+## Unreleased
+
+### Added
+
+* NeoTrace tool (#178)
+
+### Changed
+
+* Update dependencies for Neo 3.0.3 release
+
 ## [3.0.13] 2021-09-08
 
 ### Documentation